// supabase/functions/_shared/content.d.ts
type AbilityBlockType =
  | 'action'
  | 'feat'
  | 'physical-feature'
  | 'sense'
  | 'class-feature'
  | 'heritage'
  | 'mode';
type ContentType =
  | 'trait'
  | 'item'
  | 'spell'
  | 'class'
  | 'archetype'
  | 'versatile-heritage'
  | 'class-archetype'
  | 'ability-block'
  | 'creature'
  | 'ancestry'
  | 'background'
  | 'language'
  | 'content-source';
// ...
interface ContentSource {
  id: number;
  created_at: string;
  updated_at?: string;
  name: string;
  foundry_id?: string;
  url: string;
  description: string;
  operations?: Operation[];
  user_id: string;
  contact_info: string;
  require_key: boolean;
  keys?: {
    access_key?: string;
  };
  is_published: boolean;
  required_content_sources: number[];
  group: string;
  artwork_url?: string;
  meta_data?: {
    counts?: Record<ContentType | AbilityBlockType, number>;
  };
}